requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pandas
numpy
scikit-learn
matplotlib
seaborn
tqdm
opencv-python
pillow
